Extract results from the bkmr RDA files

Extract results from the bkmr RDA files

extract_data_from_folder(folder_location, select_iterations, estimate_sequence)

Arguments

folder_location

input folder location that contain bkmr output in RDA format

select_iterations

select iterations among the total number of iterations

estimate_sequence

input sequence and intervals to display overall estimates.

Value

this function will return a list of two objects the first is PIPs and second is overall risk estimates